+++ src/entry_point_win.cpp
@@ -12,7 +12,7 @@
 #include "Objbase.h"
 #include "window.hpp"
 
-#include "network/webui.hpp"
+#include "network/webapi/controllers.hpp"
 
 #pragma comment(lib, "Ole32.lib")
 #pragma comment(lib, "Shell32.lib")
